+++ CHANGELOG.md
@@ -16,7 +16,9 @@
 - Added a 3-bullet spray and cooldown pattern.
 - Added line of sight detection to trigger gunfire.
 - Added a fuse for bombs which is set upon collision rather than immediately exploding.
-- Added a level selection menu.
+- Added a level selection menu with pointer, gamepad and keyboard input.
+- Added quick restart input for gamepad, mobile and keyboard.
+- Added return to level select input for gamepad, mobile and keyboard.
 
 ## Removed
 ## Fixed